+++ src/client/pages/dashboard/dashboard.vue
