+++ src/client/router.ts
@@ -1,7 +1,7 @@
 import { createRouter, createWebHashHistory } from "vue-router";
 import Test from "./pages/test.vue";
-import Dashboard from "./pages/dashboard/index.vue";
-import Project from "./pages/project/index.vue";
+import Dashboard from "./pages/dashboard/dashboard.vue";
+import Project from "./pages/project/project.vue";
 
 const routes = [
   { path: "/", component: Dashboard },
